chore (refs T22614): move store of map to client/js.
