+++ client/js/bundles/map/mapAdminGislayerList.js
@@ -16,7 +16,7 @@ import DpSplitButton from '@DpJs/components/core/DpSplitButton'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import layers from '@DemosPlanMapBundle/store/Layers'
+import layers from '@DpJs/store/map/Layers'
 
 const stores = { layers }
 const components = {
